requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
httpx>=0.28.1
Pillow>=10.0.0

